requirements.txt
anyio==3.6.2
appdirs==1.4.4
apt-xapian-index==0.49
asgiref==3.6.0
attrs==22.2.0
beniget==0.4.1
blinker==1.5
Brlapi==0.8.4
Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
deprecation==2.0.7
distro==1.8.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dtrx==8.4.0
fastjsonschema==2.16.3
fonttools==4.38.0
fs==2.4.16
future==0.18.2
gast==0.5.2
gattlib==0.20210616
GDAL==3.6.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
idna==3.3
img2pdf==0.4.4
Jinja2==3.1.2
jsonschema==4.10.3
jupyter_core==4.12.0
kiwisolver==0.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
louis==3.24.0
lxml==4.9.2
lz4==4.0.2+dfsg
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mpmath==0.0.0
nala==0.15.4
nbformat==5.5.0
notify2==0.3
numpy==1.24.2
oauthlib==3.2.2
OWSLib==0.27.2
packaging==23.0
pbr==5.10.0
pdfarranger==1.10.0
pexpect==4.8.0
pikepdf==6.0.0+dfsg
Pillow==9.4.0
plotly==5.4.0
ply==3.11
psycopg2==2.9.5
ptyprocess==0.7.0
PyBluez==0.23
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.4.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==0.21.0
pythran==0.11.0
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
QScintilla==2.13.3
reportlab==3.6.12
requests==2.28.1
rfc3986==1.5.0
rich==13.3.1
rpl==1.14
scipy==1.10.1
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
sympy==1.11.1
tenacity==8.2.1
texttable==1.6.7
tomli==2.0.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.4.0
ufoLib2==0.14.0
ufw==0.36.2
urllib3==1.26.12
uvicorn==0.17.6
vboxapi==1.0
wadllib==1.3.6
websocket-client==1.2.3
wsproto==1.2.0
xdg==5
